requirements.txt
requests>=2.28.0
Pillow>=10.0.0
PyYAML>=6.0
streamlit>=1.28.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
# Auto text-erase batch (`src/auto_text_erase_preprocess.py`): also install paddlepaddle for your OS/CPU/GPU.
paddleocr>=2.7.0
